Adds Mailkey to UserMenu Moves UserMenu to `/components/accounts` Snapshots for UserMenu
