+++ src/lib/components/layouts/MainLayout.svelte
@@ -11,7 +11,7 @@
   import Button from "../common/Button.svelte";
   import Flex from "../common/Flex.svelte";
   import SignedIn from "../common/SignedIn.svelte";
-  import UserMenu from "../navigation/UserMenu.svelte";
+  import UserMenu from "../accounts/UserMenu.svelte";
   import OrgMenu from "../navigation/OrgMenu.svelte";
 
   import { SIGN_IN_URL } from "@/config/config";
