+++ src/lib/components/layouts/Portal.svelte
@@ -24,6 +24,6 @@
   })
 </script>
 
-<div bind:this={ref}>
+<div bind:this={ref} style="display:contents;">
   <slot />
 </div>
